Slice 2c: correct Channex ARI shapes against verified API + minor units
Validated the connector against the Channex staging sandbox (real API key)
and fixed what was provisional:

- Restrictions are per RATE PLAN: drop room_type_id from the /restrictions
  value (was wrongly included); keep {property_id, rate_plan_id, date, rate}.
- Rate must be an integer in MINOR units. Added MinorUnits converter using
  the currency's ISO-4217 fraction digits; AriUpdate.rate is now a long.
  The SPI returns Money so currency is known at the boundary. Confirmed
  empirically: IDR has 2 fraction digits, so 500000 IDR → 50000000 minor;
  a Channex round-trip reads it back as "500000.00" (sending the raw 500000
  would have under-priced inventory 100x).
- Skip past dates (Channex rejects them).
- docker-compose passes CHANNEX_* through to the app so the stack can run
  live against staging (key stays in gitignored .env).

Verified end-to-end against Channex staging: a push through our app
(outbox -> relay -> Channex) lands as availability=4 / rate=600000.00 on
read-back. Full 669-test suite green.

+++ hotel-backend/channel/src/main/java/id/co/hospitops/channel/domain/model/AriUpdate.java
@@ -1,17 +1,19 @@
 package id.co.hospitops.channel.domain.model;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
  * One night of availability + rate for a provider room type / rate plan.
  * The external ids are the provider's (Channex) identifiers, resolved from
  * the room-type mapping before the message is enqueued.
+ *
+ * <p>{@code rate} is in integer MINOR units (e.g. cents) — the form Channex
+ * accepts on ARI writes. See {@code MinorUnits}.
  */
 public record AriUpdate(
         String externalRoomTypeId,
         String externalRatePlanId,
         LocalDate date,
         int availability,
-        BigDecimal rate) {
+        long rate) {
 }
